stage4 - added Stage4UnitTest.kt

+++ Music Player/Stage 4/src/test/java/org/hyperskill/musicplayer/internals/SongFake.kt
@@ -4,5 +4,5 @@ data class SongFake(
     val id: String,
     val artist: String,
     val title: String,
-    val duration: Long
+    val duration: Int
 )
